Keyboard v1.7.1 local application finish

+++ Keylogger V1.7/zones.h
@@ -2,7 +2,7 @@
 # define _ZONES_H
 
 # include <Windows.h>
-# include <stdio.h>
+# include <Stdio.h>
 
 # include "structs.h"
 
@@ -23,7 +23,7 @@ short	z4_min(int key_value);
 short	z4_maj(int key_value);
 
 short	zone5(int key_value, keyboard_struct kb_status);
-int		z5_supp(int key_value);
+short	z5_supp(int key_value);
 wchar_t	*z5_str_add(wchar_t *src, wchar_t *str);
 
 short	zone6(int key_value, keyboard_struct kb_status);
